Alphabetical order of custom widget icons #4664
